Moved blade code into old files
