+++ src/resources/views/acctNew.php
@@ -1,11 +1,7 @@
-
-<?php
+{{-- 
 //TODO this isn't needed if we convert to Routes and Blade views
 use Illuminate\Support\ViewErrorBag;
-require_once __DIR__ . '/crsf.php'; 
-$errors = session()->get('errors', app(ViewErrorBag::class));
-//TODO this isn't needed if we convert to Routes and Blade views
-?>
+--}}
 
 <!doctype html>
 <html lang="en" data-bs-theme="auto">
@@ -13,26 +9,26 @@ $errors = session()->get('errors', app(ViewErrorBag::class));
     	<meta charset="utf-8">
     	<meta name="viewport" content="width=device-width, initial-scale=1">
     	<title>SoVest Account Creation</title>
-		<link href="css/bootstrap.min.css" rel="stylesheet">  
+		<link href="{{ asset('css/bootstrap.min.css') }}" rel="stylesheet">  
 		
-		<link rel="apple-touch-icon" sizes="180x180" href="images/apple-touch-icon.png">
-		<link rel="icon" type="image/png" sizes="32x32" href="images/favicon-16x16.png">
-		<link rel="icon" type="image/png" sizes="16x16" href="images/favicon-16x16.png">
-		<link rel="manifest" href="images/site.webmanifest">		
+		<link rel="apple-touch-icon" sizes="180x180" href="{{ asset('images/apple-touch-icon.png') }}">
+		<link rel="icon" type="image/png" sizes="32x32" href="{{ asset('images/favicon-16x16.png') }}">
+		<link rel="icon" type="image/png" sizes="16x16" href="{{ asset('images/favicon-16x16.png') }}">
+		<link rel="manifest" href="{{ asset('images/site.webmanifest') }}">		
   	</head>
   	<body>
 
 		<div class="container py-3">
   			<header>
    				<div class="d-flex flex-column flex-md-row align-items-center pb-3 mb-4 border-bottom">
-      				<a href="index.php" class="d-flex align-items-center link-body-emphasis text-decoration-none">
+      				<a href="{{ url('/') }}" class="d-flex align-items-center link-body-emphasis text-decoration-none">
        					<span class="fs-4">SoVest</span>
 					</a>
 
       				<nav class="d-inline-flex mt-2 mt-md-0 ms-md-auto">
-					  <a class="me-3 py-2 link-body-emphasis text-decoration-none" href="home.php">Home</a>
-	  					<a class="me-3 py-2 link-body-emphasis text-decoration-none" href="about.php">About SoVest</a>
-	  					<a class="me-3 py-2 link-body-emphasis text-decoration-none" href="index.php">Log In</a>
+					  <a class="me-3 py-2 link-body-emphasis text-decoration-none" href="{{ url('/home') }}">Home</a>
+	  					<a class="me-3 py-2 link-body-emphasis text-decoration-none" href="{{ url('/about') }}">About SoVest</a>
+	  					<a class="me-3 py-2 link-body-emphasis text-decoration-none" href="{{ url('/') }}">Log In</a>
       				</nav>
     			</div>
 
@@ -54,88 +50,81 @@ $errors = session()->get('errors', app(ViewErrorBag::class));
           					<div class="card-body">
 
 							<!-- Displaying Errors -->
-							<?php if ($errors && $errors instanceof MessageBag && $errors->any()): ?>
+							@if ($errors && $errors instanceof \Illuminate\Support\MessageBag && $errors->any())
 								<div class="alert alert-danger">
-								    <?php foreach ($errors->all() as $error): ?>
-										<div><?php echo htmlspecialchars($error); ?></div>
-									<?php endforeach; ?>
+								    @foreach ($errors->all() as $error)
+										<div>{{ $error }}</div>
+									@endforeach
 								</div>
-							<?php endif; ?>
-							<?php if ($errors && $errors instanceof MessageBag && $errors->any()): ?>
+							@endif
+							@if (!empty($error))
 								<div class="alert alert-danger">
-								    <?php foreach ($errors->all() as $error): ?>
-										<div><?php echo htmlspecialchars($error); ?></div>
-									<?php endforeach; ?>
-								</div>
-							<?php endif; ?>
-							  <?php if (!empty($error)): ?>
-								<div class="alert alert-danger">
-									<?php if ($error === 'invalid_email'): ?>
+									@if ($error === 'invalid_email')
 										Please enter a valid email address.
-									<?php elseif ($error === 'password_too_short'): ?>
+									@elseif ($error === 'password_too_short')
 										Password must be at least 6 characters long.
-									<?php elseif ($error === 'validation_failed'): ?>
+									@elseif ($error === 'validation_failed')
 										Please check your information and try again.
-									<?php elseif ($error === 'system_error'): ?>
+									@elseif ($error === 'system_error')
 										A system error occurred. Please try again later.
-									<?php else: ?>
+									@else
 										An error occurred. Please try again.
-									<?php endif; ?>
+									@endif
 								</div>
-							<?php endif; ?>
+							@endif
 
 								<!-- CUSTOMIZE THIS SECTION WITH FORM INFO -->
 
 								<form action="/register/submit" method="post">
-									<input type="hidden" name="_token" value="<?php echo $token; ?>">
+									@csrf
 									<div class="form-floating">
-     						 		    <input type="email" class="form-control" id="newEmail" name="newEmail" required>
- 						      	 		  <label for="newEmail">Email</label>
- 								 </div>
-   								 <br>
+     							    <input type="email" class="form-control" id="newEmail" name="newEmail" required>
+ 						      	 	  <label for="newEmail">Email</label>
+ 									 </div>
+   									 <br>
 
 								<div class="form-floating">
-    							      <input type="password" class="form-control" id="newPass" name="newPass" required>
-    							      <label for="newPass">Password</label>
-      							  </div>
-   								<br>
+    								      <input type="password" class="form-control" id="newPass" name="newPass" required>
+    								      <label for="newPass">Password</label>
+      								  </div>
+   									<br>
 
 								<div class="form-floating">
-      								  <select class="form-control" id="newMajor" name="newMajor">
+      									  <select class="form-control" id="newMajor" name="newMajor">
 										<option value="animalScience">Animal Science</option>
 									    <option value="business">Business</option>
 										<option value="computerScience">Computer Science</option>
 										<option value="creativeTechnologies">Creative Technologies</option>
-      								    <option value="education">Education</option>
-      								    <option value="economics">Economics</option>
-      								    <option value="humanities">Humanities</option>
-     								   </select>
-    								  <label for="newMajor">Current Major</label>
- 								   </div>
- 								   <br>
+      									    <option value="education">Education</option>
+      									    <option value="economics">Economics</option>
+      									    <option value="humanities">Humanities</option>
+     									   </select>
+    									  <label for="newMajor">Current Major</label>
+ 									   </div>
+ 									   <br>
 		
 
 
 								<div class="form-floating">
-      								  <select class="form-control" id="newYear" name="newYear">
+      									  <select class="form-control" id="newYear" name="newYear">
 										<option value="freshman">Freshman</option>
 									    <option value="sophomore">Sophomore</option>
 										<option value="junior">Junior</option>
 										<option value="senior">Senior</option>
-     								   </select>
-    								  <label for="newYear">Current Grade Level</label>
- 								   </div>
- 								   <br>
+     									   </select>
+    									  <label for="newYear">Current Grade Level</label>
+ 									   </div>
+ 									   <br>
 
 							
 									<div class="form-floating">
-      								  <select class="form-control" id="newScholarship" name="newScholarship">
+      									  <select class="form-control" id="newScholarship" name="newScholarship">
 										<option value="yes">Yes</option>
 									    <option value="no">No</option>
-     								   </select>
-    								  <label for="newScholarship">Are you the recipient of a work-based scholarship?</label>
- 								   </div>
- 								   <br>
+     									   </select>
+    									  <label for="newScholarship">Are you the recipient of a work-based scholarship?</label>
+ 									   </div>
+ 									   <br>
 
 									<button class="btn btn-primary w-100 py-2" type="submit">Submit</button>
 								</form>
@@ -160,6 +149,6 @@ $errors = session()->get('errors', app(ViewErrorBag::class));
 			</footer>
  
 		</div>
-		<script src="js/bootstrap.bundle.min.js"></script>
+		<script src="{{ asset('js/bootstrap.bundle.min.js') }}"></script>
 	</body>
 </html>
