+++ src/resources/views/home.php
@@ -1,25 +1,3 @@
-<?php
-session_start(); // This should be the FIRST line of the file, before any output
-
-         
-    // Retrieve the userID cookie. If not set, redirect the user to the login page. If it is set, save it as $userID
-	if(!isset($_COOKIE["userID"])){header("Location: login.php");}
-	else {$userID = $_COOKIE["userID"];}
-
-	$servername = "localhost";
-    $username = "hackberr_399";
-    $email = 'nthayslett@gmail.com';
-    $password = "MarthaBerry!";
-    $dbname = "hackberr_399";
-    $conn = mysqli_connect($servername, $username, $password, $dbname);
-    $query = "SELECT * from npedigoUser WHERE email = '{$email}'";
-    $result = mysqli_query($conn, $query);
-    $row = mysqli_fetch_assoc($result);
-    $userID = $row['id'];
-    if (!$conn) {die("Connection failed: " . mysqli_connect_error());}
-
-?>
-
 <!doctype html>
 
 <html lang="en">
@@ -27,8 +5,8 @@ session_start(); // This should be the FIRST line of the file, before any output
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>SoVest</title>
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css">
-    <link rel="icon" type="image/png" sizes="16x16" href="favicon-16x16.png">
+    <link rel="stylesheet" href="{{ asset('css/bootstrap.min.css') }}">
+    <link rel="icon" type="image/png" sizes="16x16" href="{{ asset('images/favicon-16x16.png') }}">
     <style>
         body {
             background-color: #2c2c2c;
@@ -58,19 +36,19 @@ session_start(); // This should be the FIRST line of the file, before any output
 <body>
     <nav class="navbar navbar-expand-lg navbar-dark">
         <div class="container">
-            <a class="navbar-brand" href="index.php">SoVest</a>
+            <a class="navbar-brand" href="{{ url('/') }}">SoVest</a>
             <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav">
                 <span class="navbar-toggler-icon"></span>
             </button>
             <div class="collapse navbar-collapse" id="navbarNav">
                 <ul class="navbar-nav ms-auto">
-                    <li class="nav-item"><a class="nav-link" href="search.php">Search</a></li>
-                    <li class="nav-item"><a class="nav-link" href="trending.php">Trending</a></li>
-                    <?php if ($userID): ?>
-                        <li class="nav-item"><a class="nav-link" href="account.php">My Account</a></li>
-                   <?php else: ?>
-                        <li class="nav-item"><a class="nav-link" href="index.php">Login</a></li>
-                    <?php endif; ?>
+                    <li class="nav-item"><a class="nav-link" href="{{ url('search') }}">Search</a></li>
+                    <li class="nav-item"><a class="nav-link" href="{{ url('trending') }}">Trending</a></li>
+                    @if (Auth::check())
+                        <li class="nav-item"><a class="nav-link" href="{{ url('account') }}">My Account</a></li>
+                    @else
+                        <li class="nav-item"><a class="nav-link" href="{{ url('/') }}">Login</a></li>
+                    @endif
                 </ul>
             </div>
         </div>
@@ -80,12 +58,12 @@ session_start(); // This should be the FIRST line of the file, before any output
         <h1>Welcome to SoVest</h1>
         <p>Analyze, Predict, and Improve Your Market Insights</p>
         <div class="d-flex justify-content-center gap-3 mt-4">
-            <a href="search.php" class="btn btn-primary">Search Stocks</a> 
-            <a href="trending.php" class="btn btn-warning">Trending Predictions</a>
-            <a href="<?php echo isset($userID) ? 'account.php' : 'login.php'; ?>" class="btn btn-success">My Account</a>
+            <a href="{{ url('search') }}" class="btn btn-primary">Search Stocks</a> 
+            <a href="{{ url('trending') }}" class="btn btn-warning">Trending Predictions</a>
+            <a href="{{ Auth::check() ? url('account') : url('login') }}" class="btn btn-success">My Account</a>
         </div>
     </div>
 
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
+    <script src="{{ asset('js/bootstrap.bundle.min.js') }}"></script>
 </body>
 </html>
