+++ src/resources/views/login.php
@@ -1,19 +1,15 @@
-<?php
-require_once __DIR__ . '/crsf.php'; 
-?>
-
 <!doctype html>
 <html lang="en" data-bs-theme="auto">
 	<head>
     	<meta charset="utf-8">
     	<meta name="viewport" content="width=device-width, initial-scale=1">
     	<title>SoVest</title>
-		<link href="css/bootstrap.min.css" rel="stylesheet">  
+		<link href="{{ asset('css/bootstrap.min.css') }}" rel="stylesheet">  
 		
-		<link rel="apple-touch-icon" sizes="180x180" href="images/apple-touch-icon.png">
-		<link rel="icon" type="image/png" sizes="32x32" href="images/favicon-16x16.png">
-		<link rel="icon" type="image/png" sizes="16x16" href="images/favicon-16x16.png">
-		<link rel="manifest" href="images/site.webmanifest">	
+		<link rel="apple-touch-icon" sizes="180x180" href="{{ asset('images/apple-touch-icon.png') }}">
+		<link rel="icon" type="image/png" sizes="32x32" href="{{ asset('images/favicon-16x16.png') }}">
+		<link rel="icon" type="image/png" sizes="16x16" href="{{ asset('images/favicon-16x16.png') }}">
+		<link rel="manifest" href="{{ asset('images/site.webmanifest') }}">	
 			
   	</head>
   	<body>
@@ -21,14 +17,14 @@ require_once __DIR__ . '/crsf.php';
 		<div class="container py-3">
   			<header>
    				<div class="d-flex flex-column flex-md-row align-items-center pb-3 mb-4 border-bottom">
-      				<a href="index.php" class="d-flex align-items-center link-body-emphasis text-decoration-none">
+      				<a href="{{ url('/') }}" class="d-flex align-items-center link-body-emphasis text-decoration-none">
        					<span class="fs-4">SoVest</span>
 					</a>
 
       				<nav class="d-inline-flex mt-2 mt-md-0 ms-md-auto">
-					  <a class="me-3 py-2 link-body-emphasis text-decoration-none" href="home.php">Home</a>
-	  					<a class="me-3 py-2 link-body-emphasis text-decoration-none" href="about.php">About SoVest</a>
-	  					<a class="me-3 py-2 link-body-emphasis text-decoration-none" href="login.php">Log In</a>
+					  <a class="me-3 py-2 link-body-emphasis text-decoration-none" href="{{ url('home') }}">Home</a>
+	  					<a class="me-3 py-2 link-body-emphasis text-decoration-none" href="{{ url('about') }}">About SoVest</a>
+	  					<a class="me-3 py-2 link-body-emphasis text-decoration-none" href="{{ url('login') }}">Log In</a>
       				</nav>
     			</div>
 
@@ -38,7 +34,6 @@ require_once __DIR__ . '/crsf.php';
  		 	</header>
 
 			<main>
-
 
 				<div class="row row-cols-1 row-cols-md-1 mb-1 text-center">
 
@@ -52,7 +47,7 @@ require_once __DIR__ . '/crsf.php';
 								<!-- CUSTOMIZE THIS SECTION WITH FORM INFO -->
 
 								<form action="/login/submit" method="post">
-								<input type="hidden" name="_token" value="<?php echo $token; ?>">
+								@csrf
 
 								<div class="form-floating">
        						   <input type="email" class="form-control" id="email" name="email" required>
@@ -62,9 +57,9 @@ require_once __DIR__ . '/crsf.php';
 
     <!-- SAMPLE PASSWORD FORM (WITH REQUIRED) -->
    								 <div class="form-floating">
-    						      <input type="password" class="form-control" id="password" name="password" required>
-    						      <label for="password">Password</label>
-    							    </div>
+    								      <input type="password" class="form-control" id="password" name="password" required>
+    								      <label for="password">Password</label>
+    									    </div>
    								 <br>
 
 									<button class="btn btn-success w-100 py-2" type="submit">Submit</button>
@@ -90,6 +85,6 @@ require_once __DIR__ . '/crsf.php';
 			</footer>
  
 		</div>
-		<script src="js/bootstrap.bundle.min.js"></script>
+		<script src="{{ asset('js/bootstrap.bundle.min.js') }}"></script>
 	</body>
 </html>
