--- src/resources/views/loginCheck.blade.php
@@ -1,12 +1,0 @@
-{{-- 
-Original PHP logic has been removed and should be moved to a controller.
-This file should not contain direct database access or authentication logic.
-Instead, authentication should use Laravel's built-in Auth system.
-
-Original file contained:
-- Email/password extraction from POST
-- Database connection and query
-- User validation and authentication
-- Cookie setting
-- Page redirection
---}}
